Migrate config to TS
