+++ src/lib/hooks/use-url.jsx
@@ -1,14 +1,17 @@
-import { DEFAULT_LANGUAGE, DEFAULT_SORT_TAG } from '@/lib/utils/config';
-import { createContext, useContext, useEffect, useState } from 'react';
-
 import { request } from '@/lib/api/request';
+import {
+  DEFAULT_LANGUAGE,
+  DEFAULT_ORDERING_TAG,
+  DEFAULT_SORTING_TAG,
+} from '@/lib/utils/config';
+import { createContext, useContext, useEffect, useState } from 'react';
 
 export const UrlContext = createContext();
 
 export default function UseUrlProvider({ children }) {
   const [language, setLanguage] = useState(DEFAULT_LANGUAGE);
-  const [sort, setSort] = useState(DEFAULT_SORT_TAG);
-  const [order, setOrder] = useState('desc');
+  const [order, setOrder] = useState(DEFAULT_ORDERING_TAG);
+  const [sort, setSort] = useState(DEFAULT_SORTING_TAG);
 
   const [page, setPage] = useState(1);
   const url = request.searchIssues(language, page, sort, order);
@@ -23,7 +26,17 @@ export default function UseUrlProvider({ children }) {
 
   return (
     <UrlContext.Provider
-      value={{ url, changePage, language, setLanguage, page, sort, setSort,order, setOrder }}
+      value={{
+        url,
+        changePage,
+        language,
+        setLanguage,
+        page,
+        sort,
+        setSort,
+        order,
+        setOrder,
+      }}
     >
       {children}
     </UrlContext.Provider>
