--- src/lib/utils/config.js
@@ -1,20 +1,0 @@
-export const BASE_URL = 'https://api.github.com';
-export const ISSUE_URL = `${BASE_URL}/search/issues`;
-export const LABELS = ['easy', 'first', 'good'];
-export const QUERIES = [''];
-export const LANGUAGES = [
-  'javascript',
-  'css',
-  'html',
-  'typescript',
-  'python',
-  'java',
-  'all',
-];
-export const DEFAULT_LANGUAGE = 'all';
-export const SORT_TAGS = ['comments', 'reactions', 'updated', 'best-match'];
-export const DEFAULT_SORT_TAG = 'best-match';
-export const ISSUE_PER_PAGE = 10;
-export const MAX_ISSUES_ALLOWED = 1000;
-export const TIMEOUT_SEC = 10;
-export const TOTAL_SIBLING_BUTTONS = 2;
